initial commit
The program works by asking the user to input the security name, trade type i.e. buy/sell, volume of shares to be purchased / sold, price of purchase/sale. The program then monitors the current price and completes the execution instruction when the condition is met (Buy price >= Current price or Sell price<=Current price). After successful buy/sell instruction the trader is notified and a trading summary file is created/appended
For the purpose of this exercise, have used only one security in the trading universe
Ideally would have fetched the real time value of stock prices from web, through a finance API or using JSoup. But due to compatibility issues with my Java IDE, have instead used a text file as an input to fetch current price(dummy values) of the stock. Please find text file (PriceSource.txt) which needs to be provided for the program to run. These values will be run through with a delay of 5 sec trying to replicate share price movement. The file needs to be saved in 'D' drive
The share price is assumed to change between $49-$51 for testing during the exercise